[139] 将 订单表/订单详情表 主键改为雪花算法生成而不是自动递增

+++ server/applet-server/src/main/java/com/imengli/appletServer/dao/OrderInfoDetailRepostory.java
@@ -47,5 +47,5 @@ public interface OrderInfoDetailRepostory {
     Integer save(@Param("orders") List<OrderInfoDetailDO> orders);
 
     @Select("SELECT * FROM " + SysConstant.ORDER_INFO_DEATIL_TABLE_NAME + " WHERE orderId = #{orderId}")
-    List<OrderInfoDetailDO> select(@Param("orderId") Integer id);
+    List<OrderInfoDetailDO> select(@Param("orderId") Long id);
 }
